+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.0-beta.10 [UNRELEASED]
 ### Added
 - Human/AI readable descriptors for types
-- Free run for LTC inputs
+- Free wheel for LTC inputs
 - Reference items now show a warning if they're missing a target id
 ### Fixed
 - An issue where invalid types could crash the rundown widget